remove: mandatory 'number' field from templates
Mandating the use of 'number' adds an unnecessary burden. Since this
needs to be globally unique it will conflict with not just any other
exam added that uses this plugin but *any* card that has a field named
accordingly. Sorting by question is consistent with default behaviour.
I also removed the TTS from this small section as it didn't appear
to be respecting the "false" setting.